requirements.txt
absl-py==0.7.1
alabaster==0.7.12
astor==0.7.1
audioread==2.1.6
Babel==2.6.0
boto3==1.9.156
botocore==1.12.156
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
cloudpickle==1.1.1
colorama==0.4.1
configparser==3.7.4
cycler==0.10.0
databricks-cli==0.8.6
decorator==4.3.0
dill==0.2.8.2
docker==4.0.1
docutils==0.14
EMD-signal==0.2.6
entrypoints==0.3
Flask==1.0.3
gast==0.2.2
gitdb2==2.0.5
GitPython==2.1.11
grpcio==1.20.0
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
imagesize==1.1.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
jsonschema==2.6.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
librosa==0.6.3
llvmlite==0.27.0
Markdown==3.1
MarkupSafe==1.1.0
matplotlib==2.2.3
mleap==0.8.1
mlflow==0.9.1
mock==2.0.0
multiprocess==0.70.6.1
nbformat==4.4.0
nose==1.3.7
nose-exclude==0.5.0
numba==0.42.0
numpy==1.14.5
numpydoc==0.8.0
packaging==18.0
pandas==0.24.2
pathos==0.2.2.1
pbr==5.1.3
pox==0.2.4
ppft==1.6.4.8
protobuf==3.7.1
PTable==0.9.2
pycparser==2.19
Pygments==2.3.1
pyhht==0.1.0
pyparsing==2.2.0
pyperclip==1.7.0
pypiwin32==223
pyreadline==2.1
python-dateutil==2.7.3
python-speech-features==0.6
pytz==2018.5
pywin32==224
PyYAML==5.1
querystring-parser==1.2.3
requests==2.21.0
resampy==0.2.1
retrying==1.3.3
s3transfer==0.2.0
scikit-learn==0.20.2
scipy==1.1.0
simplejson==3.16.0
six==1.11.0
smmap2==2.0.5
snowballstemmer==1.2.1
sounddevice==0.3.12
Sphinx==1.8.3
sphinx-gallery==0.2.0
sphinxcontrib-websupport==1.1.0
sqlparse==0.3.0
tabulate==0.8.3
tensorboard==1.10.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.24.1
websocket-client==0.56.0
Werkzeug==0.15.2
